Inception | Your mind is the scene of the crime
The Instructions

Brainfuck is tiny. It consists of eight different instructions. These
instructions can be used to manipulate the state of the Brainfuck
machine:

* `>` - Increment the data pointer by 1.
* `<` - Decrement the data pointer by 1.
* `+` - Increment the value in the current cell (the cell the data
pointer is pointing to).
* `-` - Decrement the value in the current cell.
* `.` - Take the integer in the current cell, treat it as an ASCII char
and print it on the output stream.
* `,` - Read a character from the input stream, convert it to an
integer and save it to the current cell.
* `[` - This always needs to come with a matching `]`. If the current
cell contains a zero, set the instruction pointer to the index of the
instruction after the matching `]`.
* `]` - If the current cell does not contain a zero, set the
instruction pointer to the index of the instruction after the matching
`[`.

That’s all of it, the complete Brainfuck language. Let's do this!

LIFE IS MAGIC, CODING IS ART.